refactored js code for threads
